+++ docs/changelog.rst
@@ -55,10 +55,19 @@ Added
     * ``reply_count``, count of replies (including replies on shares)
     * ``shares_count``, count of shares
 
+* Make email notifications nicer by using HTML templates in addition to the plain text version. (`#206 <https://github.com/jaywink/socialhome/issues/206>`_)
+
+  In addition to reply and follow notifications, send also when own content is shared.
+
 Changed
 .......
 
 * **Breaking change**. Content API results now return ``visibility`` as a string ('public', 'limited', 'site' or 'self'), not an integer.
+
+Fixed
+.....
+
+* There was no notification sent out when a local user followed a local user. This has now been fixed.
 
 Removed
 .......